+++ lib/scr/core/di/cities_injection_container.dart
@@ -2,6 +2,7 @@ import 'package:get_it/get_it.dart';
 import 'package:weather_app_dart/scr/core/api/api.dart';
 import 'package:weather_app_dart/scr/core/api/extensions/api_modal_params.dart';
 import 'package:weather_app_dart/scr/core/data/data_source/base_data_source.dart';
+import 'package:weather_app_dart/scr/core/data/data_source/cities_data_source.dart';
 import 'package:weather_app_dart/scr/core/di/external_injection_container.dart';
 
 final slCities = GetIt.instance;
