+++ lib/scr/features/city_search_page/city_search_page.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 
 import 'package:weather_app_dart/scr/core/assets/app_images.dart';
 import 'package:weather_app_dart/scr/core/data/data_source/base_data_source.dart';
+import 'package:weather_app_dart/scr/core/data/data_source/cities_data_source.dart';
 import 'package:weather_app_dart/scr/core/data/models/data_cities_model.dart';
 import 'package:weather_app_dart/scr/core/di/cities_injection_container.dart';
 
@@ -42,29 +43,6 @@ class _CitySearchPageState extends State<CitySearchPage> {
                     listData = (await citySearcDataSource
                             .fetchData(additionalQueryParams: {'q': inputCity}))
                         .listCitiesDataModel;
-                    listData.forEach(
-                      (element) => logDebug(element.cityName),
-                    );
-
-                    //new request
-                    // final url = Uri.http('dataservice.accuweather.com',
-                    //     'locations/v1/cities/autocomplete', {
-                    //   'apikey': 'LWFUAhBA7AAmO9xIv0WCUOB9s6jmWr6f',
-                    //   'q': inputCity
-                    // });
-                    // try {
-                    //   var response = await http.get(url);
-                    //   if (response.statusCode == 200) {
-                    //     final List jsonArray =
-                    //         convert.jsonDecode(response.body);
-                    //     listData.clear();
-                    //     for (var element in jsonArray) {
-                    //       listData.add(DataCityModel.fromJson(element));
-                    //     }
-                    //   }
-                    // } catch (error) {
-                    //   // todo add exception error
-                    // }
                     setState(() {});
                   }
                 },
